Initial commit of the flashcard application
This is the first pass of the flashcard application. Given a file that
alternates between a line with an English word and a line with its
Japanese meaning, the program will prompt the user to enter in the
corresponding English word.

zodiac.js contains the core code for parsing an input file and
manipulating the HTML file to be like a flashcard.

keyboard.js is visual representation of a QWERTY keyboard backed by the
HTML5 Canvas.