Redraw the keyboard if the tab loses focus
It is common for users to use Ctrl+Shift+Tab to move between tabs in a
browser. This means the shift key is depressed and is coloured red.
However, the key up event will not be received by Zodiac as the tab has
already changed. The added visibility listener will now force the
keyboard to be redrawn so that the keyboard will always be in a clean
state when the user navigates between Zodiac and another tab.