Allow more than one characters to be removed
Changed the alphabet interface so that the user can define how many
characters should be removed. The user can now also see the original
incomplete word and the word that the user is currently completing.
Correct characters are highlighted in green and incorrect characters
are highlighted as red. The user can use the backspace key to go back
if desired.