Set the background colour of the navbar's cell
The navbar currently takes up less space than the content on Chrome.
This causes the colouring to be cut off as the bullet list's rendering
is complete but the extraneous space is not rendered. The fix is to
assign a background colour to the individal cell so that the whole cell
will have that colour.