Add audio support to notify the user of right and wrong answers
When the user enters in the correct answer, a confirmation sound is
played. When the user enters an incorrect character, an error sound is
played. As long as the characters are still incorrect, the sound will
not play again. The error sound will only play when the entered
characters transition from a correct state to an incorrect state.