Fix bug that caused keys to be highlighted during a rerun
Key events were being ignored after a time trial was cut off. If the
user was pressing a key right when the time was up, the key up event
would be ignored. This meant that the keyboard would remain in a state
wherein it would think a key was still being pressed when it actually
wasn't.