Fix bug that caused a run to be successful when it failed
As the remaining time is decremented from a timer, there were cases
where a run would be successful even though it actually exceeded the
specified time of the time trial. Fix this by evaluating the success of
a time trial based on the elapsed time and the time limit instead of
the remaining number of seconds.