Implement the ability to restart the setup process
Currently, the only way to setup Zodiac again is to refresh the page.
It should be possible for the user to not only restart a run, but to
also modify the run or to perform a completely different run. To
provide this functionality, a setup button has been added to the report
so the user can now reconfigure the run without being forced to refresh
the page.