Merge the timer code together into generic functions
Both the starting countdown code and the timer code share a lot of
similarities outside of a couple of parameters for tweaking purposes.
To ensure that the maintenance of the code does not grow out of hand,
it is better to merge the two code bases so that any improvements or
bug fixes on one of the canvases will automatically be replicated onto
the other.