Optimize the reloading process by not processing the files again
When a user opts to restart a run, the original word list should simply
be cloned. The current approach of rereading the files is a source of
unnecessary file I/O. Since the original list of words is readily
available, they can simply be cloned and reused instead of rereading
the files again.